angle_min, angle_max -> float

+++ my_src/my_cfg/parser.py
@@ -49,8 +49,8 @@ class Parser:
         self.parser.add_argument('--scale_min', default=1.0, type=float, help='scale min')
         self.parser.add_argument('--scale_max', default=1.0, type=float, help='scale max')
         self.parser.add_argument('--isotropically_rescaled', action='store_true')
-        self.parser.add_argument('--angle_min', default=0, type=int, help='angle min')
-        self.parser.add_argument('--angle_max', default=0, type=int, help='angle max')
+        self.parser.add_argument('--angle_min', default=0, type=float, help='angle min')
+        self.parser.add_argument('--angle_max', default=0, type=float, help='angle max')
 
     def parse_args(self):
         parsed_args = self.parser.parse_args()
